Break apart the code some more
